refactor: 세션을 관리하는 session 디렉토리 생성
