+++ src/test/java/webserver/session/SessionTest.java
@@ -2,6 +2,7 @@ package webserver;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.session.Session;
 
 import java.lang.reflect.Field;
